requirements.txt
flask
flask-login
flask-sqlalchemy
flask-mysqldb
flask-wtf
flask-talisman
python-dotenv
deep-translator